requirements.txt
google-generativeai
google-genai
python-dotenv
Pillow
openpyxl
streamlit
pandas
